extract a slider in Testimonials to a separate component TestimonialSlider
