+++ src/components/Testimonials/TestimonialsSlider.js
@@ -1,0 +1,86 @@
+import React from 'react';
+
+import Slider from 'react-slick';
+import 'slick-carousel/slick/slick.css';
+import 'slick-carousel/slick/slick-theme.css';
+
+const TestimonialsSlider = (props) => {
+  function SampleNextArrow(props) {
+    const { className, style, onClick } = props;
+    return (
+      <div
+        className={className}
+        style={{ ...style, display: 'none', background: 'red' }}
+        onClick={onClick}
+      />
+    );
+  }
+
+  function SamplePrevArrow(props) {
+    const { className, style, onClick } = props;
+    return (
+      <div
+        className={className}
+        style={{ ...style, display: 'none', background: 'green' }}
+        onClick={onClick}
+      />
+    );
+  }
+
+  const settings = {
+    dots: true,
+    infinite: true,
+    speed: 8000,
+    slidesToShow: 3,
+    slidesToScroll: 1,
+    initialSlide: 0,
+    autoplay: true,
+
+    autoplaySpeed: 8000,
+    cssEase: 'linear',
+
+    appendDots: (dots) => (
+      <div
+        style={{
+          backgroundColor: 'white',
+          borderRadius: '10px',
+          padding: '20px',
+        }}
+      >
+        <ul style={{ margin: '0px' }}> {dots} </ul>
+      </div>
+    ),
+    responsive: [
+      {
+        breakpoint: 1574,
+        settings: {
+          slidesToShow: 2,
+          slidesToScroll: 2,
+          // infinite: true,
+          dots: true,
+        },
+      },
+      {
+        breakpoint: 600,
+        settings: {
+          slidesToShow: 2,
+          slidesToScroll: 2,
+          // initialSlide: 2,
+        },
+      },
+      {
+        breakpoint: 480,
+        settings: {
+          slidesToShow: 1,
+          slidesToScroll: 1,
+          nextArrow: <SampleNextArrow />,
+          prevArrow: <SamplePrevArrow />,
+        },
+      },
+    ],
+  };
+
+  return <Slider {...settings}> {props.children}</Slider>;
+};
+
+export default TestimonialsSlider;
